% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdim_embeddings.R
\name{clustering}
\alias{clustering}
\alias{clustering,DGrowthR-method}
\title{Clustering Function}
\usage{
clustering(object, ...)

\S4method{clustering}{DGrowthR}(
  object,
  embedding = "umap",
  algorithm = "dbscan",
  k = 0.01,
  eps = NULL
)
}
\arguments{
\item{object}{A DGrowthR object containing preprocessed data including UMAP coordinates.}

\item{...}{Additional parameters to clustering algorithm}

\item{embedding}{A string indicating if the coordinates from "umap" or "fpca" should be used for clustering.}

\item{algorithm}{A string specifying the clustering algorithm to use ("dbscan" or "gmm").}

\item{k}{Numeric value indicating the number of nearest neighbors (for DBSCAN) or number of mixture components (for GMM).
If k < 1, it is interpreted as a proportion of the total number of curves.}

\item{eps}{Numeric value indicating the epsilon distance to use for DBSCAN clustering. If NULL, automatic selection based on k is performed.}
}
\value{
Updated DGrowthR object with updated metadata dataframe indicating cluster membership.
}
\description{
This function performs clustering using the specified algorithm (DBSCAN or GMM) on UMAP or fPCA coordinates.
}
\seealso{
\code{\link[dbscan:dbscan]{dbscan::dbscan()}}, \code{\link[mclust:Mclust]{mclust::Mclust()}}
}
