% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGrowthR.R
\docType{class}
\name{DGrowthR-class}
\alias{DGrowthR-class}
\alias{DGrowthR}
\title{This S4 class represents a DGrowthR model.}
\description{
It is used to store and validate data and parameters necessary for DGrowthR analysis.
}
\section{Slots}{

\describe{
\item{\code{od_data}}{A data.frame expected to contain the following columns: "timepoint", "od", and "curve_id". These represent the time points, and optical density (od), and an identifier for each curve respectively.}

\item{\code{metadata}}{A data.frame that contains the "curve_id" column along with any metadata variables}

\item{\code{raw_od}}{Optical density information before pre-processing or re-formatting.}

\item{\code{verbose}}{A logical value indicating whether the output should be suppressed. Defaults to TRUE, i.e., output is not suppressed.}

\item{\code{fpca}}{An fpca object as the output of FPCA from fdapace}

\item{\code{umap_coord}}{Coordinates for UMAP embedding}

\item{\code{cluster_assignment}}{A data.frame matching curve_ids to the cluster assigned by clustering() function.}

\item{\code{gpfit_info}}{A list with the estimated mean gp process and the parameters used for fitting.}

\item{\code{growth_comparison}}{The results of a growth comparison.}

\item{\code{log_od}}{A logical indicating if the OD data has been logged.}

\item{\code{preprocessed}}{A logical variable indicating if the od data has been preprocessed with the preprocess_data() function.}

\item{\code{growth_parameters}}{A data.frame with the estimated growth parameters after gp modelling}
}}

\section{Validity}{

For an object to be a valid DGrowthR, it must meet the following criteria:
\itemize{
\item {The "od_data" slot must be a data frame containing the columns "timepoint", "timepoint_n", "curve_id", and "od".}
\item {The "metadata" slot must be a data frame with the "curve_id" column.}
\item {All of the "curve_id"s in the "od_data" data frame must be present in the "curve_id" column of the metadata.}
}
}

