% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.check_flat}
\alias{.check_flat}
\title{Estimate growth parameters}
\usage{
.check_flat(df)
}
\arguments{
\item{df}{A data.frame with timepoint and od fields.}
}
\value{
A logical value indicating if the curve is flat
}
\description{
Estimate growth parameters
}
\keyword{internal}
