% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.detect_phases}
\alias{.detect_phases}
\title{Iteratively detect phases
This function is heavily based on the implementation available on AMiGA.}
\usage{
.detect_phases(df, ratio_threshold, varb = "K", second_varb = "r")
}
\arguments{
\item{df}{A data.frame with the parameters of each potential growth phase.}

\item{ratio_threshold}{A numeric value indicating the minimum fraction of the carrying capacity of the first phase that the second phase must have to be considered a true growth phase.}
}
\value{
A data frame with the detected growth phases.
}
\description{
Iteratively detect phases
This function is heavily based on the implementation available on AMiGA.
}
\keyword{internal}
