% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.detect_polyauxic}
\alias{.detect_polyauxic}
\title{Detect polyauxic growth curves
This function is heavily based on the implementation available on AMiGA.}
\usage{
.detect_polyauxic(df, thresh, varb_detect)
}
\arguments{
\item{df}{A data.frame with timepoint, mean, first and second derivative fields.}

\item{thresh}{A numeric value indicating the minimum fraction of the carrying capacity of the first phase that the second phase must have to be considered a true growth phase.}

\item{varb_detect}{A character value. Either "growth_rate" or "carrying_capacity". Indicates which parameter should be used to assess the relevance of potential growth phases. Only relevant if detect_polyauxic is TRUE.}
}
\value{
A data frame.
}
\description{
Detect polyauxic growth curves
This function is heavily based on the implementation available on AMiGA.
}
\keyword{internal}
