% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.determine_auc}
\alias{.determine_auc}
\title{Determine auc}
\usage{
.determine_auc(df, tp_limit = NULL)
}
\arguments{
\item{df}{A data.frame with timepoint and mean fields.}

\item{tp_limit}{A numeric value indicating a timepoint t until which the AUC is calculated}
}
\value{
A data.frame with the empirical AUC from the mean GP fit
}
\description{
Determine auc
}
\keyword{internal}
