% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.determine_growth_loss}
\alias{.determine_growth_loss}
\title{Determine growth loss}
\usage{
.determine_growth_loss(df, max_od_data)
}
\arguments{
\item{df}{A data.frame with timepoint and mean fields.}

\item{max_od_data}{A data.frame with the information output from .determine_max_growth()}
}
\value{
A data.frame with the total loss of growth from the moment when maximum growth is reached
}
\description{
Determine growth loss
}
\keyword{internal}
