% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.determine_lag_phase}
\alias{.determine_lag_phase}
\title{Estimate lag phase}
\usage{
.determine_lag_phase(df, max_gr_data)
}
\arguments{
\item{df}{A data.frame with timepoint and second_derivative fields.}

\item{max_gr_data}{A data.frame with the output of .determine_max_growth_rate().}
}
\value{
A data.frame with the timepoint at the end of lag-phase
}
\description{
Estimate lag phase
}
\keyword{internal}
