% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.determine_max_death_rate}
\alias{.determine_max_death_rate}
\title{Estimate maximum death rate}
\usage{
.determine_max_death_rate(df, max_od_data)
}
\arguments{
\item{df}{A data.frame with timepoint and first_derivative fields.}

\item{max_od_data}{A data.frame with the information output from .determine_max_growth()}
}
\value{
A data.frame with the maximum death rate
}
\description{
Estimate maximum death rate
}
\keyword{internal}
