% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.determine_max_growth}
\alias{.determine_max_growth}
\title{Determine max growth}
\usage{
.determine_max_growth(df)
}
\arguments{
\item{df}{A data.frame with timepoint and mean fields.}
}
\value{
A data.frame with the maximum mean value and the timepoint where this occurs
}
\description{
Determine max growth
}
\keyword{internal}
