% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.determine_max_growth_rate}
\alias{.determine_max_growth_rate}
\title{Estimate maximum growth rate}
\usage{
.determine_max_growth_rate(df, max_od_data)
}
\arguments{
\item{df}{A data.frame with timepoint and first_derivative fields.}
}
\value{
A data.frame with the maximum specific growth rate and doubling time
}
\description{
Estimate maximum growth rate
}
\keyword{internal}
