% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdim_embeddings.R
\name{.determine_optimal_eps}
\alias{.determine_optimal_eps}
\title{Determine optimal value of epsilon with elbow-detection}
\usage{
.determine_optimal_eps(X, k)
}
\arguments{
\item{X}{numeric matrix of coordinates in UMAP space}

\item{k}{numeric value that indicates the K nearest neighoburs to be used}
}
\value{
numeric value corresponding to the optimal eps value
}
\description{
This function determines the optimal value of epsilon for density-based clustering based on the sorted
K-nearest neighbor distance curve. The optimal epsilon is the 'elbow' of the curve and is found as the point
with the largest distance to the (imaginary) line defined by the first and last point.
}
\keyword{internal}
