% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.determine_stationary_phase}
\alias{.determine_stationary_phase}
\title{Estimate stationary phase}
\usage{
.determine_stationary_phase(df, max_gr_data, max_od_data)
}
\arguments{
\item{df}{A data.frame with timepoint and second_derivative fields.}

\item{max_gr_data}{A data.frame with the output of .determine_max_growth_rate().}
}
\value{
A data.frame with the timepoint at the end of lag-phase
}
\description{
Estimate stationary phase
}
\keyword{internal}
