% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{.estimate_derivatives}
\alias{.estimate_derivatives}
\title{Determine derivatives}
\usage{
.estimate_derivatives(df)
}
\arguments{
\item{df}{A data.frame with timepoint and mean fields.}
}
\value{
A data.frame with the original mean data, and the first and second derivatives
}
\description{
Determine derivatives
}
\keyword{internal}
