% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_regression.R
\name{.fit_gp}
\alias{.fit_gp}
\title{Fit a Gaussian process model to data}
\usage{
.fit_gp(X_data, delete = FALSE)
}
\arguments{
\item{X_data}{A data.frame containing the covariates used for model fitting. Should have at least timepoint, timepoint_n, and od columns.}

\item{delete}{A logical value. If TRUE, the Gaussian process model will be deleted after fitting.}
}
\value{
The input object, with the Gaussian process model information as a list
}
\description{
This function fits a Gaussian process model to the provided data.
}
\keyword{internal}
