% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_regression.R
\name{.fit_lagp}
\alias{.fit_lagp}
\title{Fit a Gaussian process model to data with laGP}
\usage{
.fit_lagp(X, y, delete = FALSE)
}
\arguments{
\item{X}{A data.frame containing the covariates used for model fitting.}

\item{y}{A vector containing the values to be predicted.}

\item{delete}{A logical value. If TRUE, the Gaussian process model will be deleted after fitting.}
}
\value{
The input object, with the Gaussian process information as a list.
}
\description{
This function fits a Gaussian process model to the provided data. It is an internal function that
really only does the fit. Nothing else is done or returned by this function.
}
\keyword{internal}
