% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.merge_phases}
\alias{.merge_phases}
\title{Merge phases
This function is heavily based on the implementation available on AMiGA.}
\usage{
.merge_phases(df, idx1, idx2)
}
\arguments{
\item{df}{A data.frame with the parameters of each potential growth phase.}

\item{idx1}{row number one}

\item{idx2}{row number two}
}
\value{
A data frame with the merged growth phases.
}
\description{
Merge phases
This function is heavily based on the implementation available on AMiGA.
}
\keyword{internal}
