% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.pad_inflections}
\alias{.pad_inflections}
\title{Pad inflection points and directions
This function is heavily based on the implementation available on AMiGA.}
\usage{
.pad_inflections(ips, its, edge = 1, len = NULL)
}
\arguments{
\item{ips}{A numeric vector of inflection points (indices).}

\item{its}{A numeric vector of inflection directions (1 for positive, -1 for negative).}

\item{edge}{A numeric value indicating which edge to pad (1 for left, -1 for right).}

\item{len}{A numeric value indicating the total length of the data vector (used for right edge padding).}
}
\value{
A list containing the padded inflection points and directions.
}
\description{
Pad inflection points and directions
This function is heavily based on the implementation available on AMiGA.
}
\keyword{internal}
