% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_regression.R
\name{.predict_gp}
\alias{.predict_gp}
\title{Generate predictions from a Gaussian Process model}
\usage{
.predict_gp(
  XX,
  lagp_fit,
  complete_sigma = FALSE,
  prepare_dataframe = FALSE,
  estimate_derivatives = TRUE,
  delete = TRUE
)
}
\arguments{
\item{XX}{The covariate matrix for which we want ot make predictions}

\item{lagp_fit}{The output of an laGP fit.}

\item{complete_sigma}{A logical value indicating whether the full covariance matrix should be returned. This should be set to TRUE if sampling the posterior is desired.}

\item{estimate_derivatives}{A logical value indicating if the empirical first and second derivatives should be returned.}

\item{delete}{A logical value. If TRUE, the Gaussian Process model will be deleted from the object after prediction.}
}
\value{
The mean predicted values.
}
\description{
This function generates predictions from a Gaussian Process model fitted to an object.
The function supports "alternative" and "null" models. The "alternative" model predicts
based on both time and contrast, while the "null" model predicts based on time only.
}
\seealso{
\code{\link[laGP:predGP]{laGP::predGPsep()}}
}
\keyword{internal}
