% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_regression.R
\name{.sample_posterior_fit}
\alias{.sample_posterior_fit}
\title{Sample the posterior GP model for new curves}
\usage{
.sample_posterior_fit(lagp_fit, n_samples = 100)
}
\arguments{
\item{lagp_fit}{The output of an laGP fit.}

\item{n_samples}{A numberic value indicating the number of curves to sample from posterior}
}
\value{
A data.frame with sample_id, timepoint, and mean columns.
}
\description{
Sample the posterior GP model for new curves
}
\seealso{
\code{\link[laGP:predGP]{laGP::predGPsep()}}
}
\keyword{internal}
