% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdim_embeddings.R
\name{estimate_fpca}
\alias{estimate_fpca}
\alias{estimate_fpca,DGrowthR-method}
\title{Functional Principal Component Analysis}
\usage{
estimate_fpca(object, data = "od_data", ...)

\S4method{estimate_fpca}{DGrowthR}(object, data = "od_data", ...)
}
\arguments{
\item{object}{DGrowthR object.}

\item{data}{A string indicating the source data to embed. At the moment only embeds "od_data"}

\item{...}{Additional arguments to pass to the optns argument of FPCA}
}
\value{
The result from FPCA function from fdapace package
}
\description{
This function estimates the functional principal components of the complete dataset using the fdapace package
}
\seealso{
\code{\link[fdapace:FPCA]{fdapace::FPCA()}}
}
