% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdim_embeddings.R
\name{estimate_umap}
\alias{estimate_umap}
\alias{estimate_umap,DGrowthR-method}
\title{Functional UMAP}
\usage{
estimate_umap(object, data = "od_data", ...)

\S4method{estimate_umap}{DGrowthR}(object, data = "od_data", ...)
}
\arguments{
\item{object}{DGrowthR object.}

\item{data}{A string indicating the source data to embed. At the moment only embeds "od_data"}

\item{...}{Additional arguments to pass to the optns argument of uwot}
}
\value{
The result from umap function from uwot package
}
\description{
This function estimates the UMAP coordinates using the uwot package
}
\seealso{
\code{\link[uwot:umap]{uwot::umap()}}
}
