% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_od_data}
\alias{example_od_data}
\title{Example DGrowthR Optical Density Data}
\format{
A data frame with 10 rows and 151 columns variables:
\describe{
\item{Time}{A numerical value indicating the time at which each }
\item{Cancer_0.05_1}{Curve id}
\item{Cancer_0.05_10}{Curve id}
\item{Cancer_0.05_11}{Curve id}
\item{Cancer_0.05_12}{Curve id}
\item{Cancer_0.05_13}{Curve id}
\item{Cancer_0.05_14}{Curve id}
\item{Cancer_0.05_15}{Curve id}
\item{Cancer_0.05_16}{Curve id}
\item{Cancer_0.05_17}{Curve id}
\item{Cancer_0.05_18}{Curve id}
\item{Cancer_0.05_19}{Curve id}
\item{Cancer_0.05_2}{Curve id}
\item{Cancer_0.05_20}{Curve id}
\item{Cancer_0.05_21}{Curve id}
\item{Cancer_0.05_22}{Curve id}
\item{Cancer_0.05_23}{Curve id}
\item{Cancer_0.05_24}{Curve id}
\item{Cancer_0.05_25}{Curve id}
\item{Cancer_0.05_26}{Curve id}
\item{Cancer_0.05_27}{Curve id}
\item{Cancer_0.05_28}{Curve id}
\item{Cancer_0.05_29}{Curve id}
\item{Cancer_0.05_3}{Curve id}
\item{Cancer_0.05_30}{Curve id}
\item{Cancer_0.05_4}{Curve id}
\item{Cancer_0.05_5}{Curve id}
\item{Cancer_0.05_6}{Curve id}
\item{Cancer_0.05_7}{Curve id}
\item{Cancer_0.05_8}{Curve id}
\item{Cancer_0.05_9}{Curve id}
\item{Cancer_0.4_1}{Curve id}
\item{Cancer_0.4_10}{Curve id}
\item{Cancer_0.4_11}{Curve id}
\item{Cancer_0.4_12}{Curve id}
\item{Cancer_0.4_13}{Curve id}
\item{Cancer_0.4_14}{Curve id}
\item{Cancer_0.4_15}{Curve id}
\item{Cancer_0.4_16}{Curve id}
\item{Cancer_0.4_17}{Curve id}
\item{Cancer_0.4_18}{Curve id}
\item{Cancer_0.4_19}{Curve id}
\item{Cancer_0.4_2}{Curve id}
\item{Cancer_0.4_20}{Curve id}
\item{Cancer_0.4_21}{Curve id}
\item{Cancer_0.4_22}{Curve id}
\item{Cancer_0.4_23}{Curve id}
\item{Cancer_0.4_24}{Curve id}
\item{Cancer_0.4_25}{Curve id}
\item{Cancer_0.4_26}{Curve id}
\item{Cancer_0.4_27}{Curve id}
\item{Cancer_0.4_28}{Curve id}
\item{Cancer_0.4_29}{Curve id}
\item{Cancer_0.4_3}{Curve id}
\item{Cancer_0.4_30}{Curve id}
\item{Cancer_0.4_4}{Curve id}
\item{Cancer_0.4_5}{Curve id}
\item{Cancer_0.4_6}{Curve id}
\item{Cancer_0.4_7}{Curve id}
\item{Cancer_0.4_8}{Curve id}
\item{Cancer_0.4_9}{Curve id}
\item{DMSO_0.05_1}{Curve id}
\item{DMSO_0.05_10}{Curve id}
\item{DMSO_0.05_11}{Curve id}
\item{DMSO_0.05_12}{Curve id}
\item{DMSO_0.05_13}{Curve id}
\item{DMSO_0.05_14}{Curve id}
\item{DMSO_0.05_15}{Curve id}
\item{DMSO_0.05_16}{Curve id}
\item{DMSO_0.05_17}{Curve id}
\item{DMSO_0.05_18}{Curve id}
\item{DMSO_0.05_19}{Curve id}
\item{DMSO_0.05_2}{Curve id}
\item{DMSO_0.05_20}{Curve id}
\item{DMSO_0.05_21}{Curve id}
\item{DMSO_0.05_22}{Curve id}
\item{DMSO_0.05_23}{Curve id}
\item{DMSO_0.05_24}{Curve id}
\item{DMSO_0.05_25}{Curve id}
\item{DMSO_0.05_26}{Curve id}
\item{DMSO_0.05_27}{Curve id}
\item{DMSO_0.05_28}{Curve id}
\item{DMSO_0.05_29}{Curve id}
\item{DMSO_0.05_3}{Curve id}
\item{DMSO_0.05_30}{Curve id}
\item{DMSO_0.05_4}{Curve id}
\item{DMSO_0.05_5}{Curve id}
\item{DMSO_0.05_6}{Curve id}
\item{DMSO_0.05_7}{Curve id}
\item{DMSO_0.05_8}{Curve id}
\item{DMSO_0.05_9}{Curve id}
\item{HIV_0.2_1}{Curve id}
\item{HIV_0.2_10}{Curve id}
\item{HIV_0.2_11}{Curve id}
\item{HIV_0.2_12}{Curve id}
\item{HIV_0.2_13}{Curve id}
\item{HIV_0.2_14}{Curve id}
\item{HIV_0.2_15}{Curve id}
\item{HIV_0.2_16}{Curve id}
\item{HIV_0.2_17}{Curve id}
\item{HIV_0.2_18}{Curve id}
\item{HIV_0.2_19}{Curve id}
\item{HIV_0.2_2}{Curve id}
\item{HIV_0.2_20}{Curve id}
\item{HIV_0.2_21}{Curve id}
\item{HIV_0.2_22}{Curve id}
\item{HIV_0.2_23}{Curve id}
\item{HIV_0.2_24}{Curve id}
\item{HIV_0.2_25}{Curve id}
\item{HIV_0.2_26}{Curve id}
\item{HIV_0.2_27}{Curve id}
\item{HIV_0.2_28}{Curve id}
\item{HIV_0.2_29}{Curve id}
\item{HIV_0.2_3}{Curve id}
\item{HIV_0.2_30}{Curve id}
\item{HIV_0.2_4}{Curve id}
\item{HIV_0.2_5}{Curve id}
\item{HIV_0.2_6}{Curve id}
\item{HIV_0.2_7}{Curve id}
\item{HIV_0.2_8}{Curve id}
\item{HIV_0.2_9}{Curve id}
\item{Water_1}{Curve id}
\item{Water_10}{Curve id}
\item{Water_11}{Curve id}
\item{Water_12}{Curve id}
\item{Water_13}{Curve id}
\item{Water_14}{Curve id}
\item{Water_15}{Curve id}
\item{Water_16}{Curve id}
\item{Water_17}{Curve id}
\item{Water_18}{Curve id}
\item{Water_19}{Curve id}
\item{Water_2}{Curve id}
\item{Water_20}{Curve id}
\item{Water_21}{Curve id}
\item{Water_22}{Curve id}
\item{Water_23}{Curve id}
\item{Water_24}{Curve id}
\item{Water_25}{Curve id}
\item{Water_26}{Curve id}
\item{Water_27}{Curve id}
\item{Water_28}{Curve id}
\item{Water_29}{Curve id}
\item{Water_3}{Curve id}
\item{Water_30}{Curve id}
\item{Water_4}{Curve id}
\item{Water_5}{Curve id}
\item{Water_6}{Curve id}
\item{Water_7}{Curve id}
\item{Water_8}{Curve id}
\item{Water_9}{Curve id}
}
}
\usage{
example_od_data
}
\description{
A dataset containing measurements of optical density (OD) at various timepoints.
}
\examples{
head(DGrowthR::example_od_data)

}
\keyword{datasets}
