% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGrowthR.R
\name{growth_comparison_result}
\alias{growth_comparison_result}
\alias{growth_comparison_result,DGrowthR-method}
\title{Accessor for the result of growth comparison}
\usage{
growth_comparison_result(object)

\S4method{growth_comparison_result}{DGrowthR}(object)
}
\arguments{
\item{object}{An object from which to extract the perm_test_result slot.}
}
\value{
The growth comparison result data.frame
}
\description{
This function retrieves the result of a growth comparison
}
