% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGrowthR.R
\name{lrt_statistic}
\alias{lrt_statistic}
\alias{lrt_statistic,DGrowthR-method}
\title{Accessor for the likelihood ratio test result}
\usage{
lrt_statistic(object)

\S4method{lrt_statistic}{DGrowthR}(object)
}
\arguments{
\item{object}{An object from which to extract the perm_test_result slot.}
}
\value{
The LRT test statistic
}
\description{
This function retrieves the likelihood ratio test result from the growth comparison slot
}
