% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGrowthR.R
\name{null}
\alias{null}
\alias{null,DGrowthR-method}
\title{Accessor for the null slot}
\usage{
null(object)

\S4method{null}{DGrowthR}(object)
}
\arguments{
\item{object}{An object from which to extract the null slot.}
}
\value{
The null slot of the object.
}
\description{
This function retrieves the null slot from the provided DGrowthR object.
}
