% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGrowthR.R
\name{perm_test_result}
\alias{perm_test_result}
\alias{perm_test_result,DGrowthR-method}
\title{Accessor for the perm_test_result slot}
\usage{
perm_test_result(object)

\S4method{perm_test_result}{DGrowthR}(object)
}
\arguments{
\item{object}{An object from which to extract the perm_test_result slot.}
}
\value{
The perm_test_result slot of the object.
}
\description{
This function retrieves the perm_test_result slot from the provided DGrowthR object.
}
