% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fpca}
\alias{plot_fpca}
\alias{plot_fpca,DGrowthR-method}
\title{Plot FPCAs}
\usage{
plot_fpca(object, color = NULL)

\S4method{plot_fpca}{DGrowthR}(object, color = NULL)
}
\arguments{
\item{object}{DGrowthR object.}

\item{color}{Covariate by which to color the points in the scatterplot. Should be one of the columns in \code{metadata}. Defaults to NULL}
}
\value{
A scatterplot of the first two FPCs
}
\description{
This function plots the first two principal components calculated
}
\details{
#------------------------------------------------------------------------------------------------------------------
}
