% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_growth_comparison}
\alias{plot_growth_comparison}
\alias{plot_growth_comparison,DGrowthR-method}
\title{Plot Growth Comparison}
\usage{
plot_growth_comparison(object)

\S4method{plot_growth_comparison}{DGrowthR}(object)
}
\arguments{
\item{object}{An object of class "DGrowthR"}
}
\value{
A list with two ggplot objects. One for the null model and another one for the alternative.
}
\description{
This function plots the null and the alternative fits to the model
}
