% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_growth_curves}
\alias{plot_growth_curves}
\alias{plot_growth_curves,DGrowthR-method}
\title{Plot Optical Density}
\usage{
plot_growth_curves(object, color = NULL, facet = NULL, data = "od_data")

\S4method{plot_growth_curves}{DGrowthR}(object, color = NULL, facet = NULL, data = "od_data")
}
\arguments{
\item{object}{An object of class "DGrowthR"}

\item{color}{A column of metadata to color growth curves}

\item{facet}{A column of metadata to facet growth curves plots}

\item{data}{A character value indicating which data should be plotted. If "raw_od" then unprocessed data is plotted. If "od_data" (default) then the pre-processed data is plotted if present.}
}
\value{
A ggplot object showing lineplots of growth curves.
}
\description{
This function plots the raw optical density curves, with the possibility of specifying color and facet variables
}
