% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_od_difference}
\alias{plot_od_difference}
\alias{plot_od_difference,DGrowthR-method}
\title{Plot Optical Density over Time}
\usage{
plot_od_difference(object)

\S4method{plot_od_difference}{DGrowthR}(object)
}
\arguments{
\item{object}{An object of class "DGrowthR".}
}
\value{
A ggplot object representing the line plot.
}
\description{
Calculates the optical density between the condition and control fitted curves for each time point, plots both models as well as the optical density difference between both fits.
}
