% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_umap}
\alias{plot_umap}
\alias{plot_umap,DGrowthR-method}
\title{Plot UMAP components}
\usage{
plot_umap(object, color = NULL)

\S4method{plot_umap}{DGrowthR}(object, color = NULL)
}
\arguments{
\item{object}{Degrowth object.}

\item{color}{Covariate by which to color the points in the scatterplot. Should be one of the columns in \code{metadata}. Defaults to NULL}
}
\value{
A scatterplot of the first two UMAP components
}
\description{
This function plots the first two UMAP components calculated
}
