% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{preprocess_data}
\alias{preprocess_data}
\alias{preprocess_data,DGrowthR-method}
\title{Preprocess Growth Data}
\usage{
preprocess_data(
  object,
  baseline = 1,
  log_transform = TRUE,
  skip_first_n_timepoints = 0
)

\S4method{preprocess_data}{DGrowthR}(
  object,
  baseline = 1,
  log_transform = TRUE,
  skip_first_n_timepoints = 0
)
}
\arguments{
\item{object}{A DGrowthR object containing the growth data.}

\item{baseline}{A numeric value specifying timepoint that will be used for baseline correction. If a vector is provided, then the minimum OD of the provided timepoints is used as baseline.}

\item{log_transform}{A logical value indicating whether to perform log transformation on optical density measurements. This enables accurate estimation of growth parameters.}

\item{skip_first_n_timepoints}{A numeric value the number of timepoints to skip for all growth curves. The analysis would start at skip_first_n_timepoints + 1.}
}
\value{
The processed DGrowthR object with updated od_data.
}
\description{
This function reads growth data from a file, applies several optional preprocessing steps, and saves the processed data.
}
