% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGrowthR.R
\name{show,DGrowthR-method}
\alias{show,DGrowthR-method}
\title{Display a summary of the DGrowthR object}
\usage{
\S4method{show}{DGrowthR}(object)
}
\arguments{
\item{object}{A DGrowthR object.}
}
\description{
This method displays a summary of the DGrowthR object, including the first few rows
of the od_data slot and the log-likelihood and posterior probabilities of the alternative
and null models.
}
