% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\name{.fit_and_estimate}
\alias{.fit_and_estimate}
\title{Fit GP and gather growth parameters}
\usage{
.fit_and_estimate(
  object,
  model_covariate,
  predict_n_steps,
  downsample_every_n_timepoints,
  sample_posterior_gpfit,
  od_auc_at_t,
  sample_n_curves,
  save_gp_data,
  n_cores
)
}
\arguments{
\item{object}{A DGrowthR object containing preprocessed data.}

\item{model_covariate}{A string indicating a covariate in metadata to pool growth curves for GP modelling.}

\item{predict_n_steps}{A numeric value indicating the number of timepoints to sample from the mean posterior of the GP fit.}

\item{downsample_every_n_timepoints}{A numeric value indicating that the OD from every n timepoint should be used for GP fit. Might seep up fitting.}

\item{sample_posterior_gpfit}{A logical value. Indicates if the posterior GP should be sampled for new growth curves and parameters are estimated
from these sampled curves. Allows to estimate mean and standard deviation of growth parameters. Really only makes sense if model_covariate pools more
than one growth curve per group.}

\item{od_auc_at_t}{A numeric value indicating a specific timepoint for which the predicted OD and AUC should be returned.}

\item{sample_n_curves}{A numeric values. If sample_posterior_gpfit is TRUE, then sample_n_curves are sampled from posterior.}

\item{save_gp_data}{A logical value, indicating if the mean GP values and GP fit parameters should be saved to object.}

\item{n_cores}{A numeric values indicating the number of cores the user wants to use to model curves in parallel.}
}
\value{
Updated DGrowthR object
}
\description{
Fit GP and gather growth parameters
}
\seealso{
\code{\link[laGP:predGP]{laGP::predGPsep()}}
}
\keyword{internal}
