% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_comparison.R
\name{.gather_contrast}
\alias{.gather_contrast}
\title{Gather the data necesarry to perform a growth comparions.}
\usage{
.gather_contrast(object, contrast_vector, downsample_every_n_timepoints = 1)
}
\arguments{
\item{object}{A DGrowthR object.}

\item{contrast_vector}{A 'vector' object containing three values: 1) The variable on which the contrast is made, 2) alternative condition, and 3) reference condition}

\item{downsample_every_n_timepoints}{A numeric value indicating that the OD from every n timepoint should be used for GP fit. Might speed up fitting.}

\item{vectorA}{A vector with one of the}
}
\value{
The input object, with the appropiate OD data.

The input object, with the appropiate OD data.
}
\description{
This function gathers the relevant data for the contrast indicated by the user. It is an internal function that
is used as part of the growth_comparison method for the DGrowthR class.
}
\keyword{Compute}
\keyword{OD}
\keyword{OD||}
\keyword{This}
\keyword{across}
\keyword{all}
\keyword{between}
\keyword{computes}
\keyword{delta}
\keyword{function}
\keyword{internal}
\keyword{the}
\keyword{timepoints}
\keyword{two}
\keyword{vectors}
\keyword{||}
