% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plates.R
\name{read_individual_plate}
\alias{read_individual_plate}
\title{Read individual growth plate}
\usage{
read_individual_plate(plate_file, path_target, time_vector = c(), sep = "\\t")
}
\arguments{
\item{plate_file}{A character variable that is the file name for the plate OD data. Should contain a Time column and a column for each well.}

\item{path_target}{The path to plate_file.}

\item{time_vector}{A numeric vector that indicates the Time at which each measurement was taken.}

\item{sep}{The delimiter character that separates columns in plate_file.}
}
\value{
A data.frame with the data from plate_file. The Time column contains the values in time_vector, and the wells are renamed to be {plate_name}_{well}.
}
\description{
This function receives a file name that contains plate reader OD data. It returns a data.frame with the parsed data.
}
